requirements.txt
# ====================================
# DEPENDENCIAS DEL COTIZADOR CWS
# ====================================

# Framework web principal
Flask==3.0.0
Werkzeug==3.0.1

# Manejo de variables de entorno
python-dotenv==1.0.0

# MongoDB completamente eliminado

# Supabase PostgreSQL (UNICA BASE DE DATOS)
supabase==2.3.0
psycopg2-binary>=2.9.9

# Manejo de fechas y tiempo
python-dateutil==2.8.2

# Generación de PDFs (ReportLab es más estable)
reportlab==4.0.7

# WeasyPrint como alternativa (opcional - puede tener problemas en algunos servidores)
weasyprint==61.2

# Dependencias de WeasyPrint (se instalan automáticamente)
# cffi==1.16.0
# Pillow==10.2.0
# pycairo==1.25.1
# pyphen==0.14.0
# cssselect2==0.7.0
# html5lib==1.1
# tinycss2==1.2.1

# Utilidades adicionales
requests==2.31.0

# Servidor web para producción (REQUERIDO PARA LA NUBE)
gunicorn==21.2.0

# Google Drive API (para acceso a PDFs en la nube) - MANTENIDO COMO FALLBACK
google-api-python-client==2.144.0
google-auth==2.27.0

# Cloudinary eliminado - migrado a Supabase Storage

# Scheduler para sincronización automática
APScheduler==3.10.4

# Manejo de archivos CSV (incluido en Python estándar)
# csv - Built-in module

# Manejo de JSON (incluido en Python estándar)  
# json - Built-in module

# Otros módulos estándar necesarios:
# datetime - Built-in module
# os - Built-in module
# io - Built-in module
# atexit - Built-in module
# copy - Built-in module